projects/2602/10-youtube-shorts-agents/backend/requirements.txt
google-genai>=1.0.0
anthropic>=0.40.0
edge-tts>=6.1.0
flask>=3.0
flask-cors>=4.0
requests>=2.31
ffmpeg-python>=0.2.0
pydantic>=2.0
